Use preemptive auth if credentials specified
Without preemptive auth, the client will make two requests to the deposit URL
when basic auth is required. The first request is sent without credentials. If
the first request results in a 403 (Unauthorized) response code, a second
request is made that does include the credentials. Making two requests can
increase the likelihood of the request timing out when large files are used and
the recipient checks for authorization only after the file has been uploaded.

This commit enables preemptive authentication when any auth credentials are
supplied. With preemptive authentication, the auth credentials are sent with
the first request already, removing the need for the retry.